requirements.txt
appnope==0.1.0
astroid==1.5.3
certifi==2017.7.27.1
decorator==4.1.2
future==0.16.0
ipython==6.1.0
ipython-genutils==0.2.0
isort==4.2.15
jedi==0.10.2
lazy-object-proxy==1.3.1
mccabe==0.6.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pylint==1.7.2
python-telegram-bot==8.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
virtualenv==15.1.0
wcwidth==0.1.7
wrapt==1.10.10
